Chore: Move to compoenents layor from page
