+++ src/app/(auth)/login/reset-password/code/page.tsx
@@ -1,8 +1,8 @@
 import { Metadata } from "next";
 
 import { AuthHeaderSpotifyLogoSC } from "@/components/(auth)/serverComponents/AuthHeaderSpotifyLogoSC";
-import { ResetPasswordTitleSC } from "@/components/(auth)/serverComponents/ResetPasswordTitleSC";
-import { ResetPasswordDescriptionSC } from "@/components/(auth)/serverComponents/ResetPasswordDescriptionSC";
+import { AuthPageTitleSC } from "@/components/(auth)/serverComponents/AuthPageTitleSC";
+import { AuthPageDescriptionSC } from "@/components/(auth)/serverComponents/AuthPageDescriptionSC";
 import { AuthLinkButtonSC } from "@/components/(auth)/serverComponents/AuthLinkButtonSC";
 
 import { ResetCodeResendButtonCC } from "@/components/(auth)/clientComponents/ResetCodeResendButtonCC";
@@ -18,8 +18,8 @@ export default function ResetPasswordSendCodePage() {
   return (
     <div className="flex flex-col items-center max-w-md mx-auto text-center bg-white min-h-screen font-bold">
       <AuthHeaderSpotifyLogoSC />
-      <ResetPasswordTitleSC title="받은 편지함을 확인하세요." />
-      <ResetPasswordDescriptionSC description="이메일을 보내드렸습니다. 지침에 따라 Spotify 계정에 로그인하세요." />
+      <AuthPageTitleSC title="받은 편지함을 확인하세요." />
+      <AuthPageDescriptionSC description="이메일을 보내드렸습니다. 지침에 따라 Spotify 계정에 로그인하세요." />
       <div className="w-full">
         <AuthLinkButtonSC
           href={loginPageUrl}
