+++ src/app/(auth)/login/reset-password/email/page.tsx
@@ -1,7 +1,7 @@
 import { Metadata } from "next";
 import { AuthHeaderSpotifyLogoSC } from "@/components/(auth)/serverComponents/AuthHeaderSpotifyLogoSC";
-import { ResetPasswordTitleSC } from "@/components/(auth)/serverComponents/ResetPasswordTitleSC";
-import { ResetPasswordDescriptionSC } from "@/components/(auth)/serverComponents/ResetPasswordDescriptionSC";
+import { AuthPageTitleSC } from "@/components/(auth)/serverComponents/AuthPageTitleSC";
+import { AuthPageDescriptionSC } from "@/components/(auth)/serverComponents/AuthPageDescriptionSC";
 
 import { ResetPasswordEmailForm } from "@/components/(auth)/clientComponents/ResetPasswordEmailForm";
 
@@ -13,8 +13,8 @@ export default function ResetPasswordPage() {
   return (
     <div className="flex flex-col items-center max-w-md mx-auto text-center bg-white min-h-screen font-bold">
       <AuthHeaderSpotifyLogoSC />
-      <ResetPasswordTitleSC title="비밀번호를 재설정하세요." />
-      <ResetPasswordDescriptionSC
+      <AuthPageTitleSC title="비밀번호를 재설정하세요." />
+      <AuthPageDescriptionSC
         description="Spotify 계정에 연결된 이메일 주소를 입력해주시면 이메일을
       보내드리겠습니다."
       />
