+++ src/app/(auth)/login/reset-password/success/page.tsx
@@ -1,8 +1,8 @@
 import { Metadata } from "next";
 
 import { AuthHeaderSpotifyLogoSC } from "@/components/(auth)/serverComponents/AuthHeaderSpotifyLogoSC";
-import { ResetPasswordTitleSC } from "@/components/(auth)/serverComponents/ResetPasswordTitleSC";
-import { ResetPasswordDescriptionSC } from "@/components/(auth)/serverComponents/ResetPasswordDescriptionSC";
+import { AuthPageTitleSC } from "@/components/(auth)/serverComponents/AuthPageTitleSC";
+import { AuthPageDescriptionSC } from "@/components/(auth)/serverComponents/AuthPageDescriptionSC";
 
 import { AuthLinkButtonSC } from "@/components/(auth)/serverComponents/AuthLinkButtonSC";
 
@@ -16,8 +16,8 @@ export default function ResetPassworSuccessPage() {
   return (
     <div className="flex flex-col items-center max-w-md mx-auto text-center bg-white min-h-screen font-bold">
       <AuthHeaderSpotifyLogoSC />
-      <ResetPasswordTitleSC title="성공적으로 비밀번호가 변경되었습니다." />
-      <ResetPasswordDescriptionSC description="Spotify 계정에 로그인하세요." />
+      <AuthPageTitleSC title="성공적으로 비밀번호가 변경되었습니다." />
+      <AuthPageDescriptionSC description="Spotify 계정에 로그인하세요." />
       <div className="w-full">
         <AuthLinkButtonSC
           href={loginPageUrl}
