+++ src/app/(auth)/register/username/page.tsx
@@ -3,7 +3,7 @@ import { Metadata } from "next";
 
 import { AuthHeaderSpotifyLogoSC } from "@/components/(auth)/serverComponents/AuthHeaderSpotifyLogoSC";
 import { RegisterProgessBarAndSectionsSC } from "@/components/(auth)/serverComponents/RegisterProgessBarAndSectionsSC";
-import { BackToPreviousLevelPageButtonSC } from "@/components/(auth)/serverComponents/BackToPreviousLevelPageButtonSC";
+import { RegisterBackToPreviousPageButtonSC } from "@/app/_components/(auth)/serverComponents/RegisterBackToPreviousPageButtonSC";
 
 import { RegisterUsernameForm } from "@/app/_components/(auth)/clientComponents/RegisterUsernameForm";
 
@@ -18,7 +18,7 @@ export default function RegisterUsernamePage() {
     <div className="flex flex-col items-center max-w-md mx-auto bg-white min-h-screen">
       <AuthHeaderSpotifyLogoSC />
       <RegisterProgessBarAndSectionsSC width="66%" />
-      <BackToPreviousLevelPageButtonSC
+      <RegisterBackToPreviousPageButtonSC
         level="2"
         title="자신을 소개해주세요"
         backURL={registerPasswordUrl}
