+++ src/stories/(auth)/register/password/serverComponents.stories.tsx
@@ -1,5 +1,5 @@
 import { RegisterProgessBarAndSectionsSC } from "@/app/_components/(auth)/serverComponents/RegisterProgessBarAndSectionsSC";
-import { BackToPreviousLevelPageButtonSC } from "@/app/_components/(auth)/serverComponents/BackToPreviousLevelPageButtonSC";
+import { RegisterBackToPreviousPageButtonSC } from "@/app/_components/(auth)/serverComponents/RegisterBackToPreviousPageButtonSC";
 
 export default {
   title: "Components/Register/Password/serverComponents", // Updated title for uniqueness
@@ -14,9 +14,9 @@ export const DefaultProgessBarAndSectionsSC = {
 const registerUrl = process.env.NEXT_PUBLIC_REGISTRATION || "/register";
 
 // RegisterEmailTitleSC story
-export const DefaultBackToPreviousLevelPageButtonSC = {
+export const DefaultRegisterBackToPreviousPageButtonSC = {
   render: () => (
-    <BackToPreviousLevelPageButtonSC
+    <RegisterBackToPreviousPageButtonSC
       level="1"
       title="비밀번호를 만드세요."
       backURL={registerUrl}
